Programacion del servicio para filtrar los horarios

+++ administracion/src/main/java/com/example/administracion/Models/DiasLaborales.java
@@ -3,14 +3,17 @@ package com.example.administracion.Models;
 public enum DiasLaborales {
 
     DO (7), LU(1), MA(2), MI(3), JU(4), VI(5), SA(6);
-    private int diaLaboral;
+
+    private long diaLaboral;
     
-    private DiasLaborales(int diaLaboral) {
+    private DiasLaborales(long diaLaboral) {
         this.diaLaboral = diaLaboral;
     }
 
-    public int getDiaLaboral() {
+    public long getDiaLaboral() {
         return diaLaboral;
     }
+
+
 
 }
